final fix: ugly hack for external links
for some reason, assign domain is not correctly setting site.url +
site.baseurl and so the domain is set manually to the repo name. I’m
not sure how this will hold up when there is a custom domain name. ><
But I guess I don’t know if a lot of things will hold up with a custom
domain